Introduce NoToken authentication (WIP)
This is still a work in progress, the server side implementation isn't
there yet.

The goal is to allow a Github user to use this action without needing
to create an account or copy a token at nyrkio.com. In particular,
a contributor to a public repository, an open source project, say,
will now see their benchmark results submitted to Nyrkiö, even if they
had never heard of Nyrkiö before.