Check and force use of the same unit each time
At least rust cargo (or criterion) will report execution times
in whatever SI unit that makes the value be between 1 and 1000.
So it will happily flip-flop between 999 ns and 1.001 us.
(Of course, since it is not concerned or aware of its historical
results.)

While Nyrkiö allows you to choose the unit/magniture freely,
switching from one to another is not supported. This patch
will connect to nyrkio.com and get the previous results,
then try to convert the current unit to whatever is already
in nyrkio. Only "s" is supported, for seconds, and specifically
converting between s, ms, us or µs, and ns.